requirements.txt
numpy>=1.24
pandas>=2.0
scikit-learn>=1.3
xgboost>=2.0
pyyaml>=6.0
matplotlib>=3.7
seaborn>=0.13
fastapi>=0.115
uvicorn>=0.30
jinja2>=3.1
pyarrow>=14.0
